build: Fix Kerberos never being automatically enabled
Kerberos support was never enabled even if the support libraries were
detected on the system because the wrong variable was set. We also want
to make sure that Kerberos detection is attempted even if not requested,
and error out only if requested but not available.

We also need to make sure that "gcr" is checked for early, to avoid
kerberos being marked as available when a support library is missing.

https://bugzilla.gnome.org/show_bug.cgi?id=784513