provider: Clean up the public header
The goa_provider_ensure_credentials* APIs are only meant for the daemon
to communicate with the GoaProvider sub-classes. Therefore, there is no
need to export them as public API. They are not meant for external
consumers of libgoa-backend-1.0.so, like gnome-control-center and
gnome-initial-setup.

We can now change these interfaces without worrying about ABI/API
stability.

https://bugzilla.gnome.org/show_bug.cgi?id=784944