provider: Clean up the public header
The goa_provider_build_object API is only meant for the daemon to
communicate with the GoaProvider sub-classes. Therefore, there is no
need to export it as public API. It is not meant for external consumers
of libgoa-backend-1.0.so, like gnome-control-center and
gnome-initial-setup.

We can now change this interface without worrying about ABI/API
stability.

https://bugzilla.gnome.org/show_bug.cgi?id=784944