kerberos: Create the account only after the initial sign in
So far, a temporary account was created before attempting the initial
sign in. If successful, the password would be stored and the account
marked as permanent, otherwise it would be removed. This is different
from all the other providers, and having the provider directly write to
the secret service means that it cannot use the daemon's SecretService
instance for talking to org.freedesktop.secrets.

Instead, let's create the account only after the initial sign in has
been performed, and let the daemon add the password to the keyring.

https://bugzilla.gnome.org/show_bug.cgi?id=784944