facebook: Make it work with Graph API > 2.3
Ever since version 2.3 of the Graph API was retired on 8th July, it
stopped including "email" in the JSON when a GET request is issued
against https://graph.facebook.com/me?access_token=...

Now it needs to be explicitly asked for. The changelog [1] for version
2.4 says:
  "To try to improve performance on mobile networks, Nodes and Edges
   in v2.4 requires that you explicitly request the field(s) you need
   for your GET requests. For example, GET /v2.4/me/feed no longer
   includes likes and comments by default, but GET
   /v2.4/me/feed?fields=comments,likes will return the data."

It seems that "id" is always included, but we mention it anyway.

The "username" fallback that was added in commit 10c77f1713be2215 has
been removed because requesting it leads to:
  {
    "error": {
      "message": "(#12) username field is deprecated for versions v2.0
                  and higher",
      "type": "OAuthException",
      "code": 12,
      "fbtrace_id": "AJYNXiIv00W"
    }
  }

We will have to investigate whether we need a new fallback for the
presentation identity.

[1] https://developers.facebook.com/docs/apps/changelog

https://bugzilla.gnome.org/show_bug.cgi?id=785726