flickr, oauth: Close the dialog if the user denied access
... via the OAuth1 web page.

Ever since Flickr was added, clicking the "NO, THANKS" widget in the
web page has redirected to https://www.gnome.org/, which is specified
as GNOME's website in Flickr's App Garden, instead of to the specified
callback URI and/or returning an error. The click on this element
cannot be intercepted because the "click" event handler is not being
called (WebKit bug?). Current attempts to inspect that element breaks
the WebKit inspector.

Instead, let's also use https://www.gnome.org/ as the callback URI,
and use the lack of a query when redirecting to it to close the window.
It might need adjustments when there is a different OAuth1-based
provider.