Remove the option to preseed the providers & deprecate associated API
It's been more than four years and nobody ended up using this feature.
However, the extra cookie management that's necessary to implement this
feature has now exposed code, which people actually use, to a
WebKitGTK+ bug that's taking quite some time to be fixed. Therefore,
this is a good an excuse as any to remove this code.

This reverts commits 01447f6de49b640b7e298210aeb84390c804e9b3 and
1f714a8d536ea08fb9df0d3ce1250b4a67f35bba

https://bugzilla.gnome.org/show_bug.cgi?id=781005