identity: Silence -Wsuggest-attribute=format
The use of the custom format specifier %k trips up -Wformat. While the
GNU C library does support custom format specifiers, the simpler
alternative is to re-write set_error_from_krb5_error_code without the
use of %k. Since the format string was always used to prefix the
Kerberos error string, the loss of flexibility doesn't cost anything in
practice.

The function was renamed to reflect this change in behaviour.