kerberos: Re-write the account addition UI without realmd
So far, realmd was being merely used to auto-discover a list of realms,
and validate whether the user entered a valid realm or not before
prompting for credentials. This sounds nice in theory but, in
practice, it is far too fragile. For example, realmd gets frequently
tripped up by Fedora's Kerberos setup which uses _kerberos._tcp
instead of the more common _kerberos._udp SRV record. Secondly, the
realm validation isn't snappy and can take a while, so the user has to
wait a bit before she is able to enter her credentials.

The current UI is also odd because it presents separate fields for
entering the realm and the username. Most people are used to entering
username@realm or username@REALM because of their familiarity with
email addresses and kinit.

Therefore, the account addition UI has now been re-written without
realmd. There is only one field for the user to enter her principal,
and the realm is automatically capitalized if it isn't. It entirely
relies on libkrb5.so to validate the principal during the initial
sign-in once the credentials have been entered.

https://bugzilla.gnome.org/show_bug.cgi?id=789187