flickr: Remove unused is_deny_node implementation
Historically, Flickr has never returned an error code when the user
denied access to an account. That meant trying to parse the DOM and
detect the click on the appropriate UI element. Except, the "click"
event handler was never invoked. This might have been a WebKit bug, but
it was never diagnosed. See commit 947b7729f9d1202b.

This has gotten worse recently because attempting to inspect the UI
element breaks the WebKit inspector. There is now a workaround in place
based on having the same callback URI as the one listed as GNOME's
website in Flickr's App Garden. See commit 7f1cb7e1c47d9005.

Long story short, the is_deny_node code has never been used. It is time
to remove it.