kerberos: Add perform_initial_sign_in_finish
Having the GAsyncReadyCallback directly call g_task_propagate_boolean
on the passed in GAsyncResult leads to less code. However, the lack of
a _finish function makes it slightly more difficult to follow the logic
through a maze of similarly named functions. Plus, it doesn't hurt to
have some extra assertions and be consistent with the general pattern
of asynchronous operations in GIO.