Remove Telepathy support
The Empathy / Telepathy stack is no longer as strongly maintained as
it used to be, and the consensus in GNOME is to deprecate it.
Moreover, the Telepathy integration in GOA doesn't actually buy the
user any benefits. The original plan was to disable Empathy's own
account management UI (ie. empathy-accounts), but that never happened.
So, we ended up with two UIs, which at times had slightly divergent
behaviour - Microsoft account added via OAuth2 / XMPP versus raw
passwords / legacy pre-XMPP protocol.

These days we have Polari, which also has its own accounts UI, which
is in keeping with its status as a non-core application.

https://bugzilla.gnome.org/show_bug.cgi?id=795322