build: Remove duplicated goa_long_name
Following autotools build files, the project name was set to
`GNOME Online Accounts`, so a variable was created to be used as
package datadir and gettext domain set to `gnome-online-accounts`
string.

However, recently the project name has been changed to
`gnome-online-accounts` so the variable used as package datadir and
gettext domain is not necessary anymore.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/2