build: Remove unnecessary conditional statements
Meson's configuration object `set` method receives a boolean value.
On the other hand, there are a set of options related to the type of
accounts supported by `gnome-online-accounts`.

These options can be used as the boolean parameter directly in the
`set` method, so there is no need to use a conditional statements.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/2