Build against GLib 2.57.2
The g_clear_pointer() macro in GLib gained a bit of type safety when
using GCC and GCC-compatible compilers; now, though, GCC really does
not like it when people cast the destroy argument to GDestroyNotify.

The macro was added precisely to avoid the need to cast the arguments
for g_clear_pointer(), so the cast to GDestroyNotify was unnecessary
to begin with.