Remove Todoist and deprecate org.gnome.OnlineAccounts.Todo
During the GNOME 3.25.x / 3.26 development cycle it was decided that
Todoist doesn't fit the goals of GNOME Online Accounts [1].

The two applications that were meant to consume the Todoist provider,
GNOME Recipes [2] and To Do [3], are not part of a minimum viable GNOME
system. While they do bear the GNOME name and follow the GNOME Human
Interface Guidelines [4], they are not tightly coupled with the rest of
GNOME. eg., they don't follow the GNOME development and release
schedule. They are optional applications which often won't be
installed, and we don't want to have online accounts that don't do
anything.

Since then, Recipes has grown its own built-in Todoist support, and
To Do is expected to do the same.

[1] https://wiki.gnome.org/Projects/GnomeOnlineAccounts/Goals
[2] https://wiki.gnome.org/Apps/Recipes
[3] https://wiki.gnome.org/Apps/Todo
[4] https://developer.gnome.org/hig/stable/