httpclient: Ensure that the GTask is always cleaned up
The synchronous variant of goa_http_client_check runs a thread-specific
GMainContext until the asynchronous variant invokes its callback. Once
the callback is invoked, the GMainLoop is quit. This means that there
won't be any further iterations of the loop and no new GSources will
be dispatched.

Therefore, when the synchronous HttpClient gets aborted by the
GCancellable or due to an SSL error, then the following
http_client_check_response_cb might not be invoked due to the lack of
a running GMainLoop to drive the GMainContext; as a result, the GTask
will get leaked.

Solve this by tracking the GError and returning the GTask only when
everything is complete.

This was fixed in commit 3b35bd716dfe91ee0c1ac4b375de9dcf6e897bae, but
broke in commit 73242fa2fe5d4b18342d33b95fa449086728aa5c when porting
to GTask.

https://bugzilla.gnome.org/show_bug.cgi?id=764157