google: Only request the email field when getting the user's identity
Currently, a lot of personal information was being received, including
gender and profile picture. The email is the only thing that's really
used, for the account's identity and presentation identy, and the rest
was being simply thrown away. Therefore, avoid receiving them in the
first place.

Verified via the Google APIs Explorer and the Google OAuth 2.0
Playground:
https://developers.google.com/apis-explorer/#p/oauth2/v2/oauth2.userinfo.get
https://developers.google.com/oauthplayground/