build: Allow building with gettext ≥ 0.20
The use of AM_GNU_GETTEXT_VERSION in configure.ac instructs autopoint
to copy po/Makefile.in.in from the exact gettext version. It is fine
if the version of gettext installed on the system has the same minor
version number with the requested version, but it fails if you have a
newer version of gettext because of the mismatch between autoconf
macros and Makefile.in.in.

  *** error: gettext infrastructure mismatch: using a Makefile.in.in
  from gettext version 0.19 but the autoconf macros are from gettext
  version 0.20

Instead of specifying the exact version with AM_GNU_GETTEXT_VERSION, we
can use AM_GNU_GETTEXT_REQUIRE_VERSION to ask autopoint to simply use
the gettext version installed on the system to prevent the mismatch.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/23