provider: Reduce account widget margins
GNOME HIG recommends 18px as margin from the window edge.  Setting it
to a high value (72, 24) gives the users of goa provider less control
over the margin, making it hard to reduce the width if required.
Let’s set a smaller margin and let the users increase them if required.