kerberos-identity-manager: Try not to use GIOSchedulerJob
... and g_io_scheduler_job_send_to_mainloop.

The code is too fundamentally tied to the idea of sending a small chunk
of work from a worker thread to the main thread, and synchronously
blocking the worker thread while waiting for that work to finish.
Therefore, a private equivalent of g_io_scheduler_job_send_to_mainloop
was added that operates on GMainContext, not GIOSchedulerJob, to
unblock porting rest of the code to GTask, GThreadPool and such.

https://bugzilla.gnome.org/show_bug.cgi?id=764157