kerberos-identity-manager: Port to GThreadPool
The destruction sequence of a GoaKerberosIdentityManager instance is
simpler now because the GThreadPool makes some things easier, as
compared to having to directly handle a GAsyncQueue.

As before, there is a object-wide GCancellable and separate
GCancellables for each GAsyncResult-based cancellable action; and there
are two kinds of operations that get inserted into the thread pool -
those that are based on the GAsyncResult idiom, and those that are
internal fire-and-forget style operations. Note that due to the
semantics of the GAsyncResult idiom, a GoaKerberosIdentityManager
instance will not be destroyed while such a call is in flight because
it holds a reference to the instance.

When a GoaKerberosIdentityManager instance gets destroyed, the first
step is to cancel the object-wide GCancellable, and wait for any
ongoing to operation to complete. Then the GThreadPool gets cleared and
destroyed.

https://bugzilla.gnome.org/show_bug.cgi?id=764157