build: Decouple the identity service from the Kerberos provider
A subsequent commit will add a provider for the Fedora Account System,
which will be a branded Kerberos provider. To support enabling and
disabling the Fedora and Kerberos providers independent of each other,
it is necessary to introduce a separate variable to govern the building
of the identity service, which will be shared by the two.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/issues/26