daemon: Wipe the SessionId, if any, for permanent accounts
This can be useful when upgrading an account from temporary to
permanent, which can happen when adding any kind of Kerberos account.

The user-specified credentials of a Kerberos account can only be
verified by attempting to sign in the given principal. If successful
this will add the principal to the Kerberos credential cache. This will
create a temporary GOA account if one wasn't preemptively added before
attempting to sign in, which is what the existing Kerberos provider
does. However, this is a bit odd because it differs from the sequence
of steps taken by all the other providers.

This change will enable removing this oddity from the Kerberos provider
in future. In the more near term, a subsequent commit will add a
provider for the Fedora Account System which will be a branded
Kerberos provider, and it will take advantage of this.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/issues/26