build: Make goa-1.0.pc consistent between Autotools & Meson
The public glib-2.0 dependency was missing from the goa-1.0.pc
generated by Autotools. It doesn't matter much because gio-2.0 was
already listed and it automatically implies glib-2.0. However, it's
good to avoid needless differences between the Autotools and Meson
builds, and, as pedantic as it might be, it is more correct to specify
glib-2.0 separately.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/41