kerberos-identity-manager: Use Linux's notification pipe for KEYRING
So far, goa-identity-service didn't have a way to be notified about
changes to Kerberos credentials caches stored in the Linux kernel's
keyring. It resorted to polling the credentials caches to detect any
changes made to them, leading to higher power consumption.

The Linux kernel now offers a notification mechanism in the form of a
special pipe that can be used to listen for changes to the keyring.
This reduces the need to poll the credentials caches.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/47