daemon, dbus: Enable checking if a provider type is supported
With the advent of Flatpak, applications can no longer be expected to
be updated in lockstep with the operating system. In this case,
applications are meant to link against libgoa-1.0.so, while goa-daemon
and libgoa-backend-1.0.so are considered part of the OS.

Therefore, when a newer version of an application is running against
an older operating system, it needs to introspect the provider types
supported by the backend daemon so that features not supported by the
OS can be disabled at run-time.

Based on initial work done by Jean Felder.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/issues/43