Add GoaLinuxNotificationStream
This is a custom GUnixInputStream that's meant to be used with Linux's
notification pipes. Notifications are sent by the Linux kernel to the
read end of the pipe. The write end is unused, but should be kept open
as long as the read end is in use.

A GoaLinuxNotificationStream takes file descriptors for both the read
and write ends of a pipe, and closes the write end when the stream is
closed. This makes it easier to track the lifetime of both file
descriptors.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/merge_requests/47