data: Set up D-Bus session services as systemd user services
This allows them to be managed by `systemd --user` on systems that use
it, meaning they can be manipulated by `systemctl --user`, placed into a
slice for resource management purposes, configured with drop-in
configuration in {/etc,/usr/lib}/systemd/user/goa-daemon.service.d/*.conf
and so on.

On systems that do not launch dbus-daemon as a `systemd --user` service,
including systems that do not use systemd at all, the SystemdService
field is ignored and the dbus-daemon will continue to use the
traditional activation mechanism. In this case the GOA services will
continue to be child processes of the dbus-daemon.

Signed-off-by: Simon McVittie <smcv@collabora.com>