webview: Increase fixed size by 100×100px
At the previous fixed size of 400×500px, the "Next" button on the first
page of the Google login flow is not visible. It also feels,
subjectively, a bit cramped.

Increasing the fixed size by 100px in each axis makes the "Next" button
visible, is smaller than the default size of Initial Setup and Settings
(the parent application for the dialog this view is embedded in), and
also falls within the HIG-specified smallest display size of 1024×600.
(The existing natural_size constants already exceed the HIG-specified
minimum phone size of 360×294.)

https://developer.gnome.org/hig/guidelines/adaptive.html?highlight=size#small-size-handling