Remove Facebook and deprecate org.gnome.OnlineAccounts.Maps
Facebook was used by GNOME Maps and GNOME Photos.  However, Facebook
has been completely broken for a few years because the client code in
GNOME failed to keep up with changes in their online APIs.

This is why the Facebook support was removed from Maps in GNOME 41 [1].
Photos should do the same.

[1] https://gitlab.gnome.org/GNOME/gnome-maps/-/issues/379
    https://gitlab.gnome.org/GNOME/gnome-maps/-/merge_requests/184

https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/issues/184