kerberos-identity-manager: Clarify an ambiguous debug log about KCM
Kerberos KCM credential caches do support multiple identities and a lot
of users use KCM these days because it's the default on Fedora.  Hence,
it's better not to have a debug log that implies that the code wasn't
expecting KCM and is making assumptions about it - KCM is definitely a
supported Kerberos cache type.

https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/merge_requests/104