kerberos-identity: Clarify and remove redundancy from the renewal errors
The "Could not renew identity" segment in the error strings from
goa_kerberos_identity_renew is redundant because the caller already
knows that this function is about renewing an identity and any failure
means just that.  In fact, it's callers in GoaIdentityService and
GoaKerberosIdentityManager already prepend a similar segment when using
the error strings.

Debug logs are already verbose enough.  It's better not to make it even
more difficult to follow them through needless redundancy.

When reporting errors from krb5_cc_get_principal, it's good to make the
error string match the documentation [1] of that function because it
makes the code self-documenting.

[1] https://web.mit.edu/kerberos/krb5-devel/doc/appdev/refs/api/krb5_cc_get_principal.html

https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/issues/160