kerberos-identity: Clarify the error when talking to the KDC failed
When krb5_get_renewed_creds fails to talk to the Kerberos Key
Distribution Centre (or KDC) because of network problems,
krb5_get_error_message translates the krb5_error_code as:
  Resource temporarily unavailable

This ends up in the debug logs as:
  GoaKerberosIdentityManager: could not renew identity: Could not get
    new credentials to renew identity FOO@BAR.ORG: Resource temporarily
    unavailable
  GoaIdentityService: could not renew identity: Could not get new
    credentials to renew identity FOO@BAR.ORG: Resource temporarily
    unavailable

It's not immediately clear that the 'resource' that's 'temporarily
unavailable' is actually the KDC, which would make it easier to
understand that there is a network problem.

Therefore, mention KDC in the error string from krb5_get_renewed_creds
and make it match the documentation [1] of that function because it
makes the code self-documenting.

[1] https://web.mit.edu/kerberos/krb5-devel/doc/appdev/refs/api/krb5_get_renewed_creds.html

https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/issues/160