kerberos-identity: Unbreak handling of fresh caches
commit 4acfcc323e986526975ede981673dd173be4e267 attempted to
avoid an error variable getting stomped all over by returning
FALSE when encountering the error.

Unfortunately, it's actual legitimate for an error to happen
in that path and we should proceed anyway.

That can happen when a credential cache is new and not yet
initialized, so it won't have a principal associated with it
yet.

This commit changes the problematic code to just pass NULL
for the error variable, since we don't need it.