kerberos-identity: Fix buglet in update_identity
The update_identity function is supposed to transfer the identity
form one object to another.

In practice, this is currently always a noop because only objects
with the same identities get copied to each other.

Nevertheless, there is a bug in the function. It grabs the identity
from the target object instead of from the source object.

This commit fixes that.


(cherry picked from commit 8c5b63069dfbf36ac662369d9ba3fb4ab0d2efb2)