goaowncloudprovider: Encode user name in calendar/contacts/files URI
The GUri is constructed with a flag claiming its parts are already encoded,
but when writing verbatim identity as a user name part, it can break parsing
the GUri back from a string when the identity contains special characters
like an `@`. Writing encoded identity to such GUri fixes the problem.

Closes https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/issues/243