goakerberosidentity: Ensure properties are updated in a timely fashion
At the moment property notifications of identity objects are deferred
to a lower priority idle handler.

This is suboptimal because it means there can be a bit of a delay
updating the status of, e.g., IsSignedIn, over the bus.

This commit changes the notification to queue at normal priority.