goakerberosidentity: Queue is-signed-in notify before emitting expired signal
Right now the "expired" signal is emitted before the "is-signed-in"
property notification is queued. This notification is used to update the
"IsSignedIn" property of the object on the bus. The "expired" signal
emission leads to a corresponding "identity-expired" signal on the manager
object.  A handler for that signal calls into gnome-online-accounts to check
identity status. In response, gnome-online-accounts will check the "IsSignedIn"
property.

This commit makes sure the "is-signed-in" notification is queued first,
before the "expired" signal is emitted.

Of course queuing the notify before "expired" is emitted isn't enough to
ensure the notify happens in time. A future commit will make sure the
"identity-expired" signal itself is deferred to the main thread, which
should resolve that problem.