goakerberosidentitymanager: Ensure identity-expired signal is emitted from main loop thread
Right now most of the the identity manager signals get emitted from the
main thread. This makes sense, and is what a caller would typically
expect (given they connect their callbacks in the main thread as well).

The one exception is "identity-expired" which gets emitted from the
worker thread. This commit fixes that.