goadavclient: correct Fastmail endpoints
When Fastmail accounts are autoconfigured, the URIs are set to the
endpoint recommended in the documentation. Re-authenticating these
URIs would require the `check()` function to automagically fallback
to well-known path redirects.

Since the TXT record for fastmail.com doesn't contain a path,
amend the URIs to use the same well-known path that would be discoverd.