goaprovider: Prefer goa.conf over GSettings for enabled providers
This prefers

[providers]
enable=owncloud,kerberos

in goa.conf in the system config directory over GSettings' value,
if present, to limit which providers can be enabled. When the goa.conf
doesn't exist, or the group/key does not exist, then the value from
the GSettings is used instead.

Related to https://gitlab.gnome.org/GNOME/gnome-online-accounts/-/issues/224