goafedoraprovider: Fix possible use-after-free found by Coverity Scan
The `data->identity` had been freed, but not set to NULL, thus later in the function,
when the parse_principal() failed, there was left the `data->identity` pointing
to an already freed memory. Set it to NULL to avoid the problem.

A similar issue with `data->client_source` addressed as well.