meson: Override dependencies goa provides
This makes it easier to use wrap files to build goa from main (e.g. for
Settings), as we don't need to know the exact name of the internal meson
dependency goa provides.

See glib!1441 for more motivation in a different project.

In order for this to work, the directories 'goa' and 'goabackend' need
to be in the include path, which only works from the 'src' directory.
This is similar to how gtk declares their dependencies for example.