goabackend: don't clobber errors returned from service clients
A number of providers have been clobbering errors returned from
service clients, to the effect of making any error appear to be an
authentication error.

Double-check the clients already map errors to the GoaError domain,
and only amend the message for errors with the code already set to
`GOA_ERROR_NOT_AUTHORIZED`.